Fill in the `nvcc` compiler details
Make sure that `nvcc` is required and that both 9.2 and 10.0 versions
are used for the builds. This is added to the feedstock temporarily.
However the goal is to integrate this into conda-forge-pinning over the
long term. So that all feedstocks will benefit from the configuration
automatically and individual feedstocks merely need to use the `nvcc`
compiler as part of their build process.